Split TS generation into files : move code related to TS for links from main to other source files
