+++ src/tools/ts-generator/linksTSgenerator.h
@@ -2,6 +2,9 @@
 #pragma once
 
 #include "tsGenerationOptions.h"
+#include <antares/solver/ts-generator/generator.h>
+
+namespace fs = std::filesystem;
 
 namespace Antares::TSGenerator {
 
@@ -10,9 +13,19 @@ class LinksTSgenerator
 public:
     LinksTSgenerator(Settings&);
     void extractData();
-    void generate();
+    bool generate();
+
 private:
-    Settings settings_;
+    LinkPairs extractLinkNamesFromStudy();
+    LinkPairs extractLinkNamesFromCmdLine(const LinkPairs&);
+    StudyParamsForLinkTS readGeneralParamsForLinksTS();
+    void extractLinksSpecificTSparameters();
+
+    std::string linksFromCmdLineOptions_;
+    fs::path studyFolder_;
+    bool generateTSforAllLinks_ = false;
+    std::vector<LinkTSgenerationParams> linkList_;
+    StudyParamsForLinkTS generalParams_;
 };
 
 } // End Antares::TSGenerator
